services/orchestrator-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
httpx==0.27.0
openai==1.30.1
python-jose[cryptography]==3.3.0
redis==5.0.4
prometheus-client==0.20.0
